requirements.txt
pydantic
networkx
typer
names_generator
jinja2
hjson
python-slugify